requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.7
django-celery-results==2.5.1
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
kombu==5.5.2
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
